+++ Src/Private/Get-AbrOntapCluster.ps1
@@ -5,7 +5,7 @@ function Get-AbrOntapCluster {
     .DESCRIPTION
 
     .NOTES
-        Version:        0.6.7
+        Version:        0.6.8
         Author:         Jonathan Colon
         Twitter:        @jcolonfzenpr
         Github:         rebelinux
@@ -40,7 +40,11 @@ function Get-AbrOntapCluster {
                     'Ontap Version' = $ClusterVersion.value
                     'Number of Aggregates' = $ArrayAggr.count
                     'Number of Volumes' = $ArrayVolumes.count
-                    'Overall System Health' = $ClusterDiag.Status.ToUpper()
+                    'Overall System Health' = switch ([string]::IsNullOrEmpty($ClusterDiag.Status)) {
+                        $true { '--' }
+                        $false { $ClusterDiag.Status.ToUpper() }
+                        Default { 'Unknown' }
+                    }
                 }
                 if ($Healthcheck.Cluster.Summary) {
                     $ClusterSummary | Where-Object { $_.'Overall System Health' -like 'OK' } | Set-Style -Style OK -Property 'Overall System Health'